// frontend/src/components/admin/vigilantes/vigilantes.d.ts
export type VigilanteSimpleT = {
  id: number;
  rut: string;
  nombres: string;
  apellidos: string;
  email: string;
  rol: string;
};

export type VigilanteT = {
  id: number;
  rut: string;
  p_nombre: string;
  s_nombre: string;
  p_apellido: string;
  s_apellido: string;
  email: string;
  rol: string;
};